Refactor de componentes relacionados a las canchas. Logica en contenedor y componentes solo de renderizado
