+++ src/components/Canchas/CanchasList.jsx
@@ -3,7 +3,6 @@ import Cancha from './Cancha'
 
 const CanchasList = ({canchas}) => {
 
-  
   return (
     <div id='canchas-list'>
       {canchas.map((cancha) => <Cancha key={cancha.id} id={cancha.id} nombre={cancha.nombre} tipo={cancha.tipo} />)}
